Visualisation + buffer improvements
