+++ ios/Lift/Device/SensorDataGroupBuffer.swift
@@ -27,8 +27,8 @@ class SensorDataGroupBuffer {
     init(delegate: SensorDataGroupBufferDelegate, queue: dispatch_queue_t, deviceLocations: DeviceId -> DeviceInfo.Location) {
         self.delegate = delegate
         self.deviceLocations = deviceLocations
-        windowSize = Double(DevicePace.samplesPerPacket) / 100.0   // matches 124 samples at 100 Hz
-        windowDelay = windowSize / 2.0
+        windowSize = 1.24
+        windowDelay = 1.0
         timer = GCDTimer.createDispatchTimer(windowSize, queue: queue, block: { self.encodeWindow() })
     }
     
@@ -49,27 +49,31 @@ class SensorDataGroupBuffer {
         if let x = lastDecodeTime {
             delegate.sensorDataGroupBuffer(self, encodingSensorDataGroup: sensorDataGroup)
             
-            let start = x - windowDelay - windowSize
-            let end   = x - windowDelay
+            let start = x - windowDelay
+            let end   = x - windowDelay + windowSize
             
-            let csdas = sensorDataGroup.continuousSensorDataArrays(within: TimeRange(start: start, end: end), maximumGap: 0.3, gapValue: 0x00)
-            counter += 1
-            if !csdas.isEmpty {
-                if csdas.count > 255 { fatalError("Too many sensors") }
-                let result = NSMutableData()
-                result.appendUInt16(0xcab1)
-                result.appendUInt8(UInt8(csdas.count))
-                result.appendUInt32(counter)
-                csdas.foreach { csda in
-                    result.appendUInt16(UInt16(csda.length))
-                    let location = self.deviceLocations(csda.header.sourceDeviceId)
-                    result.appendUInt8(location.rawValue)
-                    csda.encode(mutating: result)
+            if let range = sensorDataGroup.range {
+                if range.start <= start && range.end >= end {
+                    let csdas = sensorDataGroup.continuousSensorDataArrays(within: TimeRange(start: start, end: end), maximumGap: 0.3, gapValue: 0x00)
+                    counter += 1
+                    if !csdas.isEmpty {
+                        if csdas.count > 255 { fatalError("Too many sensors") }
+                        let result = NSMutableData()
+                        result.appendUInt16(0xcab1)
+                        result.appendUInt8(UInt8(csdas.count))
+                        result.appendUInt32(counter)
+                        csdas.foreach { csda in
+                            result.appendUInt16(UInt16(csda.length))
+                            let location = self.deviceLocations(csda.header.sourceDeviceId)
+                            result.appendUInt8(location.rawValue)
+                            csda.encode(mutating: result)
+                        }
+                        
+                        delegate.sensorDataGroupBuffer(self, continuousSensorDataEncodedAt: start, data: result)
+                        sensorDataGroup.removeSensorDataArraysEndingBefore(start)
+                    }
                 }
-                
-                delegate.sensorDataGroupBuffer(self, continuousSensorDataEncodedAt: start, data: result)
-                sensorDataGroup.removeSensorDataArraysEndingBefore(start - windowSize)
-            } 
+            }
         }
         
     }
